More control over docker container output streams
We allow clients to specify one of two behaviors for each stream
descriptor: either explicitly declare "_stdout" or "_stderr" as
task outputs to be stored in memory and pushed, or stream them
to the job output log if they are not explicitly declared as
task outputs. We make use of the select() system call to watch
both streams at once in a non-blocking manner.