Remove the explicit data type for tasks
Any python type can now be used as data for a workflow.  The
idea before was to force explicit interfaces for data types, but
that idea worked out to be too rigid and impractical for reasonable
use cases.  This patch only removes the classes without replacing
the port generation methods that data types defined.