Remove broken newick to vtktree converter
see http://www.paraview.org/Bug/view.php?id=15610

The move of R converters into their own plugin caused romanesco to
pick vtktree as the intermediate format between newick and nested,
instead of r.apetree as it was doing before. This uncovered a bug
in vtkNewickTreeReader that was causing a crash, so this converter
was removed.