Integrate NetworkX into core portions of codebase
Re #111

This introduces the use of validators and converters as a NetworkX
DiGraph instead of a nested dictionary.

Type/Formats are represented as a 2 element tuple such as ('string',
'text')

The nodes of the DiGraph contain the analysis needed to validate an
input of it's own format. The edges contain the analysis needed to
convert an input going from its source validator to its target
validator.

This doesn't break any backwards compatibility (to my
knowledge). Plugins can continue to call import_converters on certain
paths and they will just be added to the DiGraph. Conversion paths are
computed at the moment they need to be used.

This removes the notion of a 3-hop cutoff. This is something that will
be noted in the future by introducing an "explain"-like operation which
will show the path different analyses will have to take.