Add debug flag, writes script to file before exec
Interactive debuggers generally choke on calls to 'exec'  with strings,
they produce output like:

(Pdb)
> <string>(1)<module>()

rather than:

(Pdb)
1: output = isinstance(input, (str, unicode))

This commit allows the user to add a 'debug' runtime flag to either an
analysis,  or to the kwargs of a call to romanesco.run().  When true,
rather than calling exec on the task script as a string,  the task
script will be written to a temporary file then exec'ed as a file handle
object. This appears to resolve most inconsistencies with interactive
debuggers and exec.

If 'debug: 1' is added to the analysis json only the analysis will be
written to file before executed.  If added to the kwargs of a call to
romanesco.run() all analyses (including conversion and validation) will
be written to temporary files.