Add AnonymousTask and Task classes and tests
This commit changes the Task class to an 'AnonymousTask' and implements a
derivative class 'Task.'  AnonymousTasks may set inputs and outputs
ports in the spec passed into the Task at the time of
instantiation. Classes that derive from Task must define the
inputs/outputs ports in the class level variables __inputs__ and
__outputs__.

This begins to make a distinction between classes and objects in the
port/spec definition,  where classes map to analyses,  and objects map
to instances of analyses that may have data associated with them.