Fix incorrect kwargs passing call
The reason we construct the kwargs in this cryptic way is because
this function is capable of recursing, and in so doing the kwargs
get passed down multiple times in the recursive calls. This
construction ensures that the task_input and task_output keys only
occur once in the kwargs; if they appear multiple times, it
triggers a runtime error.