Fix a test failure caused by Pillow upgrade
From the Pillow 2.7.0 release notes:
"The GaussianBlur() implementation has been replaced with a sequential
application of box filters."

This results in a slightly different pixel mean value after blurring.

"There was an error in the previous version of Pillow, where blur radius
(the standard deviation of Gaussian) actually meant blur diameter. For
example, to blur an image with actual radius 5 you were forced to use
value 10. This has been fixed. Now the meaning of the radius is the same
as in other software.

If you used a Gaussian blur with some radius value, you need to divide
this value by two."

This requires the input "blur_radius" parameters to be updated in the test.