Fix a test failure
One of the tests makes calls to api.opentreeoflife.org. This service seems
to now require the use of HTTPS.

Despite the fact that the Requests library is used to make these calls,
this issue appears to be unrelated to the upgrade of this library (the
failure occurs even with the prior version of the library).