Make docker gc opt-in, rather than opt-out.
If you don't add a value of "gc: True" to the docker section of the config, containers are run with the --rm flag to removed themselves when done, and the docker gc script is never run (and therefore never deletes images).  If "gc: True" is specified, then containers linger after use and docker gc will clean things up with the current rules (such as cache time and exclusion lists).