Add GW container and common_tasks GW task plugin to the mix.
Note that common_tasks.egg-info must exist and be version
controlled. Both girder & girder_worker containers create egg-link
files in /usr/local/lib/python2.7/dist-packages that point to
/girder_worker/tests/integration/common_tasks. This means if you mount
common_tasks at run time there must be an egg-info folder for the
package to be importable (and for entry points to be found).