Change default output path behavior in docker mode
Prior to this change, if no output path was specified, the
id of the output was used as the basename, relative to the
data volume mount point. Now, the default behavior is to use
$output{id}.

This commit also changes the behavior when expanding $input{}
templates to throw an exception if the input id is invalid.