Refactor decorated task discovery
The old method of task discovery relied on an internal registry
which only stored reference to bare functions.  It did not allow
one to discover which functions were requested by *each* extension.
The new code searches modules on demand for valid tasks.  In
addition, it allows the caller to filter tasks to only those
wrapped as a celery task.