Fix a version conflict for idna when installing
Currently when installing into a new virtualenv, pip will fail
with a message saying:

```
error: idna 2.6 is installed but idna<2.6,>=2.5 is required by set(['requests'])
```

This is due to requests using a `<` specifier while PyOpenSSL depends on
(indirectly) `idna>=2.1`.  Because PyOpenSSL is processed first the
version resolution fails for requests.  This should be fixed soon
but will continue to happen until requests stops pinning its
dependencies.  See the following:

https://github.com/requests/requests/issues/4222