Stop crashing when girder_worker_pypi_version isn't passed
Prior to this commit, pypi installs via the ansible role were broken
unless the user specified a version with girder_worker_pypi_version
since it was defaulting to the empty string (equivalent to pip install
girder-worker==). Since the variable is technically undefined now, the
default will be to install the latest off pypi (not specifying ==).