Add deployment step to import role into ansible-galaxy
This commit adds a deployment step to the circle ci configuration
which will run galaxy_deploy.sh on the master branch. This means that
every time the tests pass on a commit on master (merges), a push will
occur to girder/ansible-role-girder-worker and ansible galaxy will be
told to re-import the role.

In terms of security, this requires the following:
- ANSIBLE_GALAXY_GITHUB_TOKEN be defined as an environment variable in
Circle. This is a token which allows the user:read permission.
- A read-write deploy key for ansible-role-girder-worker which allows
Circle to push to this repo (Circle has the private key).