Use direct paths when possible.
If a direct_path parameter is given as part of a input specification, and that path is reachable, use it instead of downloading the file.  This only works for files where the parent is not being fetched.

This changes the girder_io plugin to use the direct path instead of download the file to the temp directory.

This changes the docker plugin to mount direct path files as their own read-only volumes.  This maintains the same level of isolation as before, except that the path name on the girder system is revealed to the worker.