Fallback to a regular container if nvidia is unavailable.
When running an nvidia container, if we can't connect to nvidia, run as a regular container.

Ideally, it would be good if we knew if nvidia was *required* or only *preferred*.  When it is just preferred, it is better to run without nvidia support then to not run at all.