Rename girder_worker/tests girder_worker/_test_plugins
This makes plugin code used for tests apart of a "private" module
inside girder worker. It is difficult to mock out this kind of code
because it relies on setuptools' entrypoint infrastructure.  That
means there need to be actual stubbed out classes/tasks in some
installed package for plugin loading tests to work.