Fix DLL references between plugins
This implementation relies on specifying dependency information via a
special key inside package.json.  Ideally we would want to avoid doing
this and automatically determine dll references and build order from the
normal dependencies (or peerDepenencies) value.  This might be possible
in the future; however, we might also want to remove DLL's entirely in
favor of using native code splitting.