Add a getLocalFilePath method to the File model.
The method `getLocalFilePath` is added to the File model, and will
return None if there is no such path.

The same method name is used in the AbstractAssetstoreAdapter and
inherited by the other adapters.  The filesystem assetstore adapter
overrides this to provide a path.  The Fuse plugin overrides the
AbstractAssetstoreAdapter to report a fuse path if one exists.

This allows any consumer (such as the worker or metadata extractor
plugins) to ask for `File().getLocalFilePath(file)`, and, if it isn't
`None`, to use that path.

If the Fuse plugin is enabled, the metadata extractor will now work on
any file.  This also fixes a bug where a file that was attached to a
resource (rather than on an Item) would throw an exception in the
metadata extractor.