Add error handling logic when running with girder 2.x
Pre-girder 3.0 plugins supported the same entrypoint (`girder.plugin`)
to register themselves as Girder 3.  If a Girder 2.x instance is
installed with this package, it will attempt to load this entrypoint and
fail with an error message.  Due to how pre-3.0 plugins worked, it still
falls back to the built-in version of worker and functions correctly.
This just removes the confusing error message.

Note: Prior to this commit, the error was only emitted when installed in
non-development mode.