Fix (hack) weird documentation issues with CLI
Prior to this PR gwrun used the default "function" documentation to
provide help for commands.  This is problematic because the function
is actual a Celery Task whose documentation comes from
celery.local.Proxy. This caused the following behavior:

Usage: gwrun [OPTIONS] COMMAND [ARGS]...

Options:
  -o, --output [json|stdout]
  --help                      Show this message and exit.

Commands:
  add              Proxy that evaluates object once.
  subtract         Proxy that evaluates object once.
  multiply         Proxy that evaluates object once.
  divide           Proxy that evaluates object once.

Now we explicitly set the help string to "" if there is no
documentation on the wrapped celery function. If there IS
documentation on the function is is passed through correctly.