Set sessions to retry connections.
As of PR #376, we are using request sessions by default.  If a job
doesn't emit output for a long time, then tries to emit output, that
session object might have been disconnected by the server, at which
point the job fails.  Allowing any retries fixes this.

Perhaps there is a better way to have both sessions and not have to add
a retry adapter.